Create clean architecture for both factory and program
