+++ src/FreelanceBotBase.Bot/Commands/Factory/Abstract/ICallbackCommandWithInputFactory.cs
@@ -1,0 +1,19 @@
+﻿using FreelanceBotBase.Bot.Commands.Interface;
+using FreelanceBotBase.Domain.State;
+
+namespace FreelanceBotBase.Bot.Commands.Factory.Abstract
+{
+    /// <summary>
+    /// Factory that creates callback commands and awaits user input.
+    /// </summary>
+    public interface ICallbackCommandWithInputFactory
+    {
+        /// <summary>
+        /// Creates callback command with input.
+        /// </summary>
+        /// <param name="inputState">Input state.</param>
+        /// <param name="chatId">Chat id.</param>
+        /// <returns></returns>
+        ICallbackCommandWithInput CreateCallbackCommandWithUserInput(BotState.InputState inputState, long chatId);
+    }
+}
